feat(wip): gui render loop

+++ arch/i686/hal/memory/memory.h
@@ -7,6 +7,6 @@
 #define SEGOFF2LIN(off) ((SEG(off) << 4) + OFF(off))
 
 
-void* memcpy(void* dst, const void* src, uint16_t num);
+void* memcpy(void* dst, const void* src, uint32_t num);
 void* memset(void* ptr, int value, uint16_t num);
 int memcmp(const void* ptr1, const void* ptr2, uint16_t num);
